final version for DBdesign and fixtures
