--- src/DataFixtures/traits/hardcodedData/AmenityHardcodedData.php
@@ -1,7 +1,0 @@
-<?php
-namespace App\DataFixtures\traits\hardcodedData;
-
-trait AmenityHardcodedData
-{
-
-}
