--- src/DataFixtures/traits/hardcodedData/ArtistHardcodedData.php
@@ -1,7 +1,0 @@
-<?php
-namespace App\DataFixtures\traits\hardcodedData;
-
-trait ArtistHardcodedData
-{
-
-}
