--- src/DataFixtures/traits/hardcodedData/ReviewHardcodedData.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\DataFixtures\traits\hardcodedData;
-
-class ReviewHardcodedData
-{
-
-}
