+++ src/DataFixtures/traits/hardcodedData/RoleData.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\DataFixtures\traits\hardcodedData;
+
+trait RoleData
+{
+    private const ROLE_NAMES = [
+        'Administrator',
+        'Super Admin',
+        'Content Manager',
+        'Event Organizer',
+        'Festival Director',
+        'Marketing Manager',
+        'Social Media Manager',
+        'Customer Support',
+        'Ticket Manager',
+        'Artist Coordinator',
+        'Venue Manager',
+        'Security Manager',
+        'Technical Support',
+        'Finance Manager',
+        'HR Manager',
+        'Volunteer Coordinator',
+        'Media Relations',
+        'Sponsorship Manager',
+        'Operations Manager',
+        'Quality Assurance',
+        'Data Analyst',
+        'Community Manager',
+        'Content Creator',
+        'Event Photographer',
+        'Sound Engineer',
+        'Stage Manager',
+        'Logistics Coordinator',
+        'Guest Relations',
+        'VIP Manager',
+        'Merchandise Manager'
+    ];
+}
