+++ src/DataFixtures/traits/hardcodedData/UserData.php
@@ -1,0 +1,22 @@
+<?php
+namespace App\DataFixtures\traits\hardcodedData;
+
+trait UserData
+{
+    private const FIRST_NAMES = [
+        'Alexandru', 'Maria', 'John', 'Wei', 'Yuki', 'Mohammed',
+        'Sofia', 'Ivan', 'Giuseppe', 'Hans', 'Fatima', 'Pedro',
+        'Raj', 'Olga', 'Lars', 'Chen', 'Aisha', 'Carlos',
+        'Emma', 'Antoine', 'Dimitri', 'Sven', 'Isabella', 'Klaus',
+        'Priya', 'Miguel', 'Andrei', 'Yusuf', 'Anna', 'Giovanni'
+    ];
+
+    private const LAST_NAMES = [
+        'Popescu', 'Smith', 'Zhang', 'Tanaka', 'Al-Sayed',
+        'García', 'Ivanov', 'Rossi', 'Mueller', 'Hassan',
+        'Silva', 'Patel', 'Kowalski', 'Andersson', 'Li',
+        'Kumar', 'Rodriguez', 'Dubois', 'Petrov', 'Bergmann',
+        'Ferrari', 'Sanchez', 'Nielsen', 'Cohen', 'Kim',
+        'Santos', 'Kovac', 'Ahmed', 'Jensen', 'Romano'
+    ];
+}
