+++ src/DataFixtures/userRelated/RoleFixtures.php
@@ -2,6 +2,7 @@
 
 namespace App\DataFixtures\userRelated;
 
+use App\DataFixtures\traits\hardcodedData\RoleData;
 use App\Entity\Role;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Bundle\FixturesBundle\FixtureGroupInterface;
@@ -10,38 +11,7 @@ use Doctrine\Persistence\ObjectManager;
 
 class RoleFixtures extends Fixture implements DependentFixtureInterface, FixtureGroupInterface
 {
-    private const ROLE_NAMES = [
-        'Administrator',
-        'Super Admin',
-        'Content Manager',
-        'Event Organizer',
-        'Festival Director',
-        'Marketing Manager',
-        'Social Media Manager',
-        'Customer Support',
-        'Ticket Manager',
-        'Artist Coordinator',
-        'Venue Manager',
-        'Security Manager',
-        'Technical Support',
-        'Finance Manager',
-        'HR Manager',
-        'Volunteer Coordinator',
-        'Media Relations',
-        'Sponsorship Manager',
-        'Operations Manager',
-        'Quality Assurance',
-        'Data Analyst',
-        'Community Manager',
-        'Content Creator',
-        'Event Photographer',
-        'Sound Engineer',
-        'Stage Manager',
-        'Logistics Coordinator',
-        'Guest Relations',
-        'VIP Manager',
-        'Merchandise Manager'
-    ];
+    use RoleData;
 
     public function load(ObjectManager $manager): void
     {
