+++ src/DataFixtures/userRelated/UserDetailsFixtures.php
@@ -3,6 +3,7 @@
 namespace App\DataFixtures\userRelated;
 
 use App\DataFixtures\traits\AppGeneralConstants;
+use App\DataFixtures\traits\hardcodedData\UserData;
 use App\Entity\UserDetails;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Bundle\FixturesBundle\FixtureGroupInterface;
@@ -10,24 +11,8 @@ use Doctrine\Persistence\ObjectManager;
 
 class UserDetailsFixtures extends Fixture implements FixtureGroupInterface
 {
-    private const FIRST_NAMES = [
-        'Alexandru', 'Maria', 'John', 'Wei', 'Yuki', 'Mohammed',
-        'Sofia', 'Ivan', 'Giuseppe', 'Hans', 'Fatima', 'Pedro',
-        'Raj', 'Olga', 'Lars', 'Chen', 'Aisha', 'Carlos',
-        'Emma', 'Antoine', 'Dimitri', 'Sven', 'Isabella', 'Klaus',
-        'Priya', 'Miguel', 'Andrei', 'Yusuf', 'Anna', 'Giovanni'
-    ];
 
-    private const LAST_NAMES = [
-        'Popescu', 'Smith', 'Zhang', 'Tanaka', 'Al-Sayed',
-        'García', 'Ivanov', 'Rossi', 'Mueller', 'Hassan',
-        'Silva', 'Patel', 'Kowalski', 'Andersson', 'Li',
-        'Kumar', 'Rodriguez', 'Dubois', 'Petrov', 'Bergmann',
-        'Ferrari', 'Sanchez', 'Nielsen', 'Cohen', 'Kim',
-        'Santos', 'Kovac', 'Ahmed', 'Jensen', 'Romano'
-    ];
-
-    use AppGeneralConstants; //constants between fixtures
+    use AppGeneralConstants, UserData;
     public function load(ObjectManager $manager): void
     {
         for ($i = 0; $i < self::TOTAL_EXISTING_USERS; $i++) {
